requirements.txt
py-cpuinfo
psutil
distro
colorama
pyfiglet